baselines/3d_btf/requirements.txt
numpy
Pillow
scikit-learn
scipy
sklearn
timm
torch
torchvision
tqdm
wget
open3d
tifffile
scikit-image
kornia
